adicionado de env, gitignore e modificado handle de error para diretorio
