+++ src/app/.env
@@ -1,0 +1,1 @@
+ MONGODB_URI=
