+++ src/app/.gitignore
@@ -1,0 +1,3 @@
+frontend/node_modules/
+server/node_modules/
+
