+++ src/app/server/controllers/authControllers.js
@@ -1,36 +1,11 @@
 const User = require("../model/authModel");
 const jwt = require("jsonwebtoken");
-
+const { handleErrors } = require('../errors/AuthenticationError');
 const maxAge = 3 * 24 * 60 * 60;
 const createToken = (id) => {
   return jwt.sign({ id }, "kishan sheth super secret key", {
     expiresIn: maxAge,
   });
-};
-
-const handleErrors = (err) => {
-  let errors = { email: "", password: "" };
-
-  if (err.message === "incorrect email") {
-    errors.email = "Email está incorreto!";
-  }
-
-  if (err.message === "incorrect password") {
-    errors.password = "Senha está incorreta!";
-  }
-
-  if (err.code === 11000) {
-    errors.email = "Email já registrado!";
-    return errors;
-  }
-
-  if (err.message.includes("Users validation failed")) {
-    Object.values(err.errors).forEach(({ properties }) => {
-      errors[properties.path] = properties.message;
-    });
-  }
-
-  return errors;
 };
 
 module.exports.register = async (req, res, next) => {
